ServerAnalysis/requirements.txt
geopy
pyproj==3.7.1
scikit-learn==1.7.1
haversine
scipy==1.16.0
numpy==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
deprecation==2.1.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.112.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
firebase-admin==6.8.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
multidict==6.6.3
packaging==25.0
passlib==1.7.4
postgrest==1.0.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
pydantic-settings==2.1.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rsa==4.9.1
sentry-sdk==2.32.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.37.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
typer==0.16.0
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
websockets==14.2
yarl==1.20.1
psycopg2-binary==2.9.10
pandas~=2.3.1
celery==5.4.0
redis==5.2.1
flower>=2.0.1
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0